+++ BufferUtils.h
@@ -18,6 +18,7 @@
 
 
 #include <Carbon/Carbon.h>
+#import <MacTypes.h>
 
 #define ResizeArray(__DirectBuffer, __objCount, __bufObjCount)	_ResizeBuffer((void***)(__DirectBuffer), (__objCount), (__bufObjCount), sizeof(typeof(**(__DirectBuffer))))
 
@@ -38,10 +39,10 @@ typedef struct _PerDiskInfo {
 } PerDiskInfo;
 
 char *replaceString(char *oldString, const char *newString);
-void _ResizeBuffer(void ***buffer, unsigned int objCount, unsigned int *bufSize, unsigned int elemSize);
+void _ResizeBuffer(void ***buffer, NSUInteger objCount, NSUInteger *bufSize, NSUInteger elemSize);
 int IsZeros(const void *s1, size_t n);
 int ContainsUInteger(const NSUInteger *uintArray, size_t count, NSUInteger auint);
-void modp_tolower_copy(char* dest, const char* str, int len);
+void modp_tolower_copy(char* dest, const char* str, NSUInteger len);
 void replace_breaks_utf8(char *s, size_t up_to_len);
 void replace_breaks(char *str, size_t up_to_len);
 int ContainsHighAscii(const void *s1, size_t n);
@@ -51,7 +52,7 @@ NSInteger genericSortContextLast(void* one, void* two, int (*context) (void*, vo
 void QuickSortBuffer(void **buffer, unsigned int objCount, int (*compar)(const void *, const void *));
 
 void RemovePerDiskInfoWithTableIndex(UInt32 diskIndex, PerDiskInfo **perDiskGroups, unsigned int *groupCount);
-unsigned int SetPerDiskInfoWithTableIndex(UTCDateTime *dateTime, UInt32 *nodeID, UInt32 diskIndex, PerDiskInfo **perDiskGroups, unsigned int *groupCount);
+NSUInteger SetPerDiskInfoWithTableIndex(UTCDateTime *dateTime, UInt32 *nodeID, UInt32 diskIndex, PerDiskInfo **perDiskGroups, NSUInteger *groupCount);
 void CopyPerDiskInfoGroupsToOrder(PerDiskInfo **flippedGroups, unsigned int *existingCount, PerDiskInfo *perDiskGroups, size_t bufferSize, int toHostOrder);
 
 CFStringRef CreateRandomizedFileName(void);
