+++ NSData_transformations.h
@@ -18,7 +18,7 @@
 - (BOOL) isCompressedFormat;
 
 + (NSMutableData *)randomDataOfLength:(int)len;
-- (NSMutableData*)derivedKeyOfLength:(int)len salt:(NSData*)salt iterations:(int)count;
+- (NSMutableData*)derivedKeyOfLength:(NSUInteger)len salt:(NSData *)salt iterations:(NSUInteger)count;
 - (unsigned long)CRC32;
 - (NSData*)SHA1Digest;
 - (NSData*)MD5Digest;
@@ -38,7 +38,7 @@
 
 @interface NSMutableData (NVCryptoRelated)
 - (void)reverseBytes;
-- (void)alignForBlockSize:(int)alignedBlockSize;
+- (void)alignForBlockSize:(NSUInteger)alignedBlockSize;
 
 - (BOOL)encryptAESDataWithKey:(NSData*)key iv:(NSData*)iv;
 - (BOOL)decryptAESDataWithKey:(NSData*)key iv:(NSData*)iv;
